Correctly handle IP on GAE vs. standalone
On App Engine the `RemoteAddr` doesn't have a port. Running
`net.SplitHostPort()` on an IPv6 address in GAE causes an error because
there are too many colons and the IPv6 address isn't surrounded by `[]`.

This change looks for an App Engine header in the request and if
present, use uses `RemoteAddr` as-is, otherwise `SplitHostPort()` is
used.