Add /info with AS and country information
This adds the /info endpoint to return a JSON object with information
about the IP address provided by the Team Cymru WHOIS service.

To try and avoid abusing the Team Cymru service, data for each IP is
cached in memcache. Memcache is only implemented on App Engine
currently. Stubs are in place for the standalone server.