Split whoisIP connection handling
App Engine requires to use their implementation of `net.Dial()` so that
it uses the request context. For some reason this isn't required in the
dev app server so only failed in the live deployment.

This moves the connection creation into a new `newWhoisConn()` function,
split into appengine and non-appengine files.