Update Docker build
Update the Dockerfile now that it's in the root of the project to copy
the whole source tree in for the build. This allows building using local
changes for testing.

The Go version is bumped to 1.12 and an /etc/passwd file is installed so
the binary can run as `nobody` instead of `root`.

`EXPOSE` is removed as we shouldn't need that and `docker run` can
specify it.

The Makefile is modified to use `docker push` directly as the `gcloud`
tool no longer needs to use its own version.