Add support for X-Forwarded-For
When deploying the server behind a proxy using the request's remote
address doesn't help as it's likely some internal IP. In these cases the
X-Forwarded-For header is added.

This adds support for parsing this header, always using the last
globally-routable address in the chain as the presumed public IP of the
client.

Some extra parsing of the IP needs to be done because Go's net library
has a function called IsGlobalUnicast() which is a bit of a lie: it
reports addresses as being global even if they're private addresses. A
reasonably comprehensive list of non-routed networks is evaluated
against the proxy address to determine the last routable address.