Replace whoisIP with iptoasn package
I created the iptoasn package (github.com/jamesog/iptoasn) to be a more
general utility library. It uses the DNS interface to Team Cymru,
whereas the implementation in this package uses the TCP WHOIS service.

DNS is generally faster and cached in various places, so I've ripped out
the memcache code from this package. Memcache was also only supported in
first-gen runtimes in App Engine; second-gen required using
externally-hosted Redis or another GCP product.

I may add caching back at a later point but for now the DNS service is
generally good enough that it's not required and shouldn't hurt Team
Cymru's service if someone decides to abuse this service.