Refactor net.go
The nonGlobalNets variable was only built at startup from init() in
net.go, however this isn't totally useful in tests because Go always
uses client IP 192.0.2.1 which is in the TEST-NET-1 documentation
prefix.

A new net_test.go file overrides this to exclude the documentation
prefixes, which also required factoring out the code to build the slice
of prefixes to its own function.