Fix tests
The tests have been broken since commit
2f4cdbb56430036e97bea14b0996a74c1dd469f6, when the IP lookup
functionality was moved to a separate package.

The tests for /info have been removed as they're not easy to mock, as
the iptoasn package uses DNS lookups instead of whois, and you can't
tell Go's DNS client code to use a specified resolver; netgo only uses
/etc/resolv.conf.

The / code now tests that XFF was parsed properly, or on Fly.io, that it
used the Fly-Client-IP header.